refactor: move setupMenu function to utils of workflow
